comm-hub/requirements.txt
psycopg2-binary==2.9.7
twilio==8.5.0
requests==2.31.0
flask==2.3.3